requirements.txt
Flask==3.0.0
flask-socketio==5.3.6
eventlet==0.33.3
python-socketio==5.8.0
python-engineio==4.7.1
opencv-python==4.9.0.80
numpy==1.26.4
requests==2.31.0